Move CrashHandler to it's own package
We want to start using DumpWithoutCrash functionality, it means that
CrashHandler should be visible at least from OrbitGl. Therefore,
CrashHandler should be moved from OrbitQt to it's own package.
