+++ CMakeLists.txt
@@ -161,6 +161,7 @@ add_subdirectory(src/OrbitVersion)
 if(WITH_GUI)
   add_subdirectory(src/CodeViewer)
   add_subdirectory(src/CodeViewerDemo)
+  add_subdirectory(src/CrashHandler)
   add_subdirectory(src/OrbitGgp)
   add_subdirectory(src/OrbitGl)
   add_subdirectory(src/OrbitSsh)
