+++ src/CrashHandler/CrashOptions.cpp.in
@@ -1,0 +1,11 @@
+// Copyright (c) 2020 The Orbit Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#include "CrashHandler/CrashOptions.h"
+
+#define CRASH_SERVER_URL "@CRASHDUMP_SERVER@"
+
+namespace orbit_crash_handler {
+std::string GetServerUrl() { return CRASH_SERVER_URL; }
+}  // namespace orbit_crash_handler
