+++ src/CrashHandler/include/CrashHandler/CrashOptions.h
@@ -1,0 +1,14 @@
+// Copyright (c) 2020 The Orbit Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#ifndef ORBIT_CRASH_HANDLER_CRASH_OPTIONS_H_
+#define ORBIT_CRASH_HANDLER_CRASH_OPTIONS_H_
+
+#include <string>
+
+namespace orbit_crash_handler {
+[[nodiscard]] std::string GetServerUrl();
+}  // namespace orbit_crash_handler
+
+#endif  // ORBIT_CRASH_HANDLER_CRASH_OPTIONS_H_
