--- src/OrbitQt/CrashOptions.cpp.in
@@ -1,11 +1,0 @@
-// Copyright (c) 2020 The Orbit Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-#include "CrashOptions.h"
-
-#define CRASH_SERVER_URL "@CRASHDUMP_SERVER@"
-
-namespace CrashServerOptions {
-std::string GetUrl() { return CRASH_SERVER_URL; }
-}  // namespace CrashServerOptions
