--- src/OrbitQt/CrashOptions.h
@@ -1,14 +1,0 @@
-// Copyright (c) 2020 The Orbit Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-#ifndef ORBIT_QT_CRASH_OPTIONS_H_
-#define ORBIT_QT_CRASH_OPTIONS_H_
-
-#include <string>
-
-namespace CrashServerOptions {
-std::string GetUrl();
-}  // namespace CrashServerOptions
-
-#endif  // ORBIT_QT_CRASH_OPTIONS_H_
